GROW-174 Reactive: Add lock repository
The db migration had to be changed to use `WITH TIME ZONE` instead of
`WITHOUT`, because otherwise the `valid_until` value was stored/retrieved
with the wrong time zone (1h off).

The lock repository is implemented slightly different compared to the
blocking/hibernate based one, but this reactive one here should be a
bit more efficient (regarding locks) and still correct as shown by
the tests.