GROW-174 Improve OutboxLockRepository, fix preventLockStealing
The implementation of `preventLockStealing` was actually wrong,
because it also returned true if the expected lock was not found.

Testing this and the `acquireOrRefreshLock` logic for cases where
a lock already exists (the "refresh" case) showed, that a lock owner
could be prevented by a lock stealing attempt from refreshing the
lock (because the lock stealing attempt directly selected the
record `FOR UPDATE`). This is now changed, and the lock owner
directly `FOR UPDATE`, which is an optimization though (to save
one additional sql query for the good/normal case which is also
the one done more frequently).