Add commons module, consolidate build setup
The new commons module provides the things that were provided by
`one.tomorrow.kafka:kafka-utils`. These are pulled into this library
because creating a separate project and standalone library doesn't seem
to be worth it, compared to the value / functionality it brings and also
considering the effort that comes with maintaining a library.

Existing code at Tomorrow could be migrated to use the
transactional-outbox:commons module. The kafka-utils module could also
be adjusted to provide a "bridge" to transactional-outbox:commons.

The build setup is simplified, i.e. a root/parent `build.gradle.kts` is
introduced which provides the common plugins and settings. This is also
a preparation for further build / publishing configuration. Former build
setup (tied to the Tomorrow internal build infrastructure) is removed
already, because it's now no longer needed/used.