Add publish job
The idea of publishing is that this happens on every merge: either to
the Sonatype snapshots repo or to the staging release repo. I.e. before
merging the version should be the right one. Releases have to be
published "manually", following https://central.sonatype.org/publish/release/

The publish job is added to the existing actions file, to have things
as close as possible. Also the build job is reused by publish, instead
of running `gradle build publish`.
Of course there are also arguments for doing it differently, I
had a slight preference for this approach though.