Set `enable.idempotence = true` by default
And rely on defaults for related properties
`max.in.flight.requests.per.connection`, `retries` and `acks`.

If the user has configured `enable.idempotence = false` a warning is
logged - assuming that the user probably knows what she's doing.

The motivation was discussed in
https://github.com/tomorrow-one/transactional-outbox/issues/39:
with `enable.idempotence = true` out-of-order messages are prevented
for the active producer (the sequence check on server side checks
if `nextSeq == lastSeq + 1L` - so that missing messages are detected).
With `enable.idempotence` we can use the default
`max.in.flight.requests.per.connection = 5` to get increased throughput.