Optimize outbox processing to enable producer batching
This was an insight in the context of issue #39: the classic module
processed outbox records sequentially, awaiting the successful
publishing of one record, updating its `processed` property, and only
then processing the next record.

This made producer internal batching impossible and reduces the
throughput. For low traffic environments this might not make a
difference, but with more traffic this might become an issue and limit
throughput too much.

Especially with PR #134, with `idempotence = true` set by the
`DefaultKafkaProducerFactory`, the
`max.in.flight.requests.per.connection` would be 5 (which was set to 1
before), which also allows more throughput even on a single partition.