Fix protobuf classpath dependency for classpath scanning
When the library is integrated together with `@ComponentScan` and
protobuf is not on the classpath, the application start failed with a
`NoClassDefFoundError` for the protobuf `Message`:

```
Caused by: org.springframework.beans.factory.BeanCreationException: Error creating bean with name 'protobufOutboxService': Lookup method resolution failed
    ...
  Caused by: java.lang.NoClassDefFoundError: com/google/protobuf/Message
    at java.base/java.lang.Class.getDeclaredMethods0(Native Method)
    at java.base/java.lang.Class.privateGetDeclaredMethods(Class.java:3402)
    at java.base/java.lang.Class.getDeclaredMethods(Class.java:2504)
    at org.springframework.util.ReflectionUtils.getDeclaredMethods(ReflectionUtils.java:465)
```

This is fixed now with a custom `Conditional` annotation on
`ProtobufOutboxService`, which prevents loading if `Message` is not
on the classpath.

While spring boot already provides a `ConditionalOnClass` annotation,
this would have pulled in a dependency on spring-boot, which would be a
big thing for spring-classic applications. Therefore, we have our own
`ConditionalOnClass` annotation now.