Replace hibernate with jdbc
For projects not using JPA/hibernate but e.g. spring-data-jdbc, the
former implementation meant to setup and configure hibernate appropriately.
Switching to plain jdbc should simplify things and make the library
easier to include. For hibernate users it should still be easy, the
`JdbcTemplate` should be available in spring-boot projects out of the
box, in classic spring it would still be easy to configure it.

Looking at artifact dependencies: spring-orm has a dependency on
spring-jdbc, i.e. we're not pulling in spring-jdbc as a new dependency,
only remove spring-orm. For json (headers) processing we're now
explicitly depending on jackson. The postgres dependency is now an
implementation dependency, because we're mapping the json headers on our
own (instead of via `hibernate-types`).