Release classic 3.2.0 and reactive 3.1.1
For the classic variant the minor version is pushed because the change
from hibernate to jdbc is considered to be more than a patch release.
Still, it's expected to be a drop-in replacement and apps *should* not
need to be adjusted.

The reactive version is also released, just for the sake of it (only the
postgres dependency should be updated).