Fix #282: Support disabled auto-commit
If an application is configured with disabled auto-commit for the
db driver / pool, the `processed` property of outbox records is not
updated after a record was sent to kafka - because the update is rolled
back instead of being is committed.

The fix is to add the `@Transactional` annotation to the
`updateProcessed` method. Note that in fact is was there when jpa was
used - because there it was definitely needed. With the transition to
spring-jdbc it seemed as if the `@Transactional` is not needed - but
this was only true for enabled auto-commit (which is the default for
postgres).

(cherry picked from commit 667359668860bd871243a0337336bfa2bbffb915)