Fix #304: Add automatic cleanup of processed outbox records
This is implemented in a backwards compatible manner, i.e. existing
integrations would have to be adjusted to use automatic cleanup.

The cleanup is done by the active instance holding the lock. It's using
a separate executor/scheduler, so that outbox processing cannot be
negatively affected / blocked by the cleanup.