Add tracing context propagation (classic)
This provides an implementation for
[micrometer-tracing](https://docs.micrometer.io/tracing/reference/), the
default tracing abstraction for Spring Boot since 3.2 (as successor for
Spring Cloud Sleuth). Micrometer Tracing provides abstractions for
different tracing solutions on it's own. If something else like e.g.
Spring Cloud Sleuth or another tracing solution is used this should
still be possible by providing an implementation for `TracingService`.

What does the current solution provide?
* Adds a span for the message in the transactional-outbox (with start
  timestamp set to the time when the outbox record was created, and end
  timestamp set to the time when it's processed and sent to Kafka)
* Adds another span for the sending the message to Kafka (start
  timestamp is the time when the `ProducerRecord` is created, end
  timestamp is set when message sending got confirmed)
* Add tracing headers to the Kafka message, which can be extracted via
  `Propagator.extract` on consumer side.

The `micrometer-tracing` dependency shall not be pulled by this library,
but would have to be added by the application. In consequence, the
`MicrometerTracingService` bean is only available, if the application is
using `micrometer-tracing`, otherwise the `NoopTracingService` is used
as fallback implementation. The current solution via `@Primary` may not
be optimal, if this should become an issue, we could add our own
`@ConditionalOnMissingBean` annotation with the appropriate handling (to
not add a dependency on spring-boot, because also plain spring
applications shall be supported). Another alternative could be to
provide proper Autoconfiguration.