feat: Integrate JReleaser for publishing to Maven Central via Sonatype Publisher Portal (#540)
This is the migration to JReleaser for publishing new SNAPSHOT and Prod releases of the library. We need to switch to another publishing mechanism regarding the end of life of OSSRH on 30.06.2025. Unfortunately, Gradle does not directly support publishing to Maven Central via the Sonatype Central Portal. The first mentioned plugin in the migration documentation was JReleaser.