Add quarkus extension - initial version
The outbox-kafka-quarkus module contains runtime functionality, the
deployment module the required build steps for quarkus. The deployment
module duplicates a few test related things from the runtime module to
keep the build simple - in general we rather copy things than making
the build more complicated.

The runtime module uses the quarkus plugin for tests so that we get
quarkus test support and are able to run integration tests. Apart from
that the runtime module shall just be built via the quarkus extension
plugin together with the deployment module.

The root gradle build file contains some exceptions for plugins that are
needed to make it work, e.g. because the protobuf plugin does not work
with quarkus (and we're not using protobuf in the quarkus module - for
now we assume it's not needed, can be added on request).

Some functionality like trace propagation still needs to be added, also
documentation is not yet there.