Add tracing context propagation to quarkus
This is very similar to the solution implemented with PR #494 (in commit
d1d38345) for spring classic, only that it's using opentelemetry, which
is the default tracing solution for quarkus.

What does the solution provide:
* Adds a span for the message in the transactional-outbox (with start
  timestamp set to the time when the outbox record was created, and end
  timestamp set to the time when it's processed and sent to Kafka)
* Adds another span for the sending the message to Kafka (start
  timestamp is the time when the `ProducerRecord` is created, end
  timestamp is set when message sending got confirmed)
* Add tracing headers to the Kafka message, which can be extracted via
  `Propagator.extract` on consumer side.

When opentelemetry is on the classpath, context propagation should be
active, otherwise there's no tracing context propagation.